backend/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
Authlib==1.3.2
bcrypt==5.0.0
blinker==1.8.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
cryptography==46.0.1
cycler==0.12.1
deprecation==2.1.0
easyocr==1.7.2
exceptiongroup==1.3.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.57.0
fsspec==2025.3.0
gotrue==2.9.2
h11==0.16.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.9
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.7.5
midtransclient==1.4.2
mpmath==1.3.0
networkx==3.1
ninja==1.13.0
numpy==1.24.4
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==10.4.0
pip==25.0.1
polars==1.8.2
postgrest==0.16.11
psutil==7.1.0
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.1.4
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyWavelets==1.4.1
PyYAML==6.0.2
realtime==1.0.6
requests==2.32.4
scikit-image==0.21.0
scipy==1.10.1
setuptools==41.2.0
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
storage3==0.7.7
StrEnum==0.4.15
supabase==2.6.0
supafunc==0.5.1
sympy==1.13.3
tifffile==2023.7.10
torch==2.4.1
torchvision==0.19.1
typing_extensions==4.13.2
ultralytics==8.3.203
ultralytics-thop==2.0.17
urllib3==2.2.3
websockets==12.0
Werkzeug==3.0.6
zipp==3.20.2
